Update custom_admin_menu.py to use view permmissions
View permissions were introduced in Django 2.1. This file has been modified to check if the version of Django is greater or equal to 2.1 and defaults to using "view" permissions when determining which models to expose in the menus. I believe that "view" perms are the least required permissions for supported versions as Django will ultimately decide how to present the model to the user based on whether "change" permission is also given to the model. "View" permission should suffice for including in the menu.